Add improvements in table styling
