+++ app/javascript/src/components/Url/HandleShortenUrl.jsx
@@ -2,13 +2,12 @@ import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import urlsApi from "apis/urls";
 
-const HandleShortenUrl = ({ fetchUrls }) => {
+const HandleShortenUrl = () => {
   const { slug } = useParams();
 
   const getOriginalLinkBySlug = async () => {
     try {
       await urlsApi.show(slug);
-      fetchUrls();
     } catch (error) {
       console.log(error);
     }
@@ -16,7 +15,7 @@ const HandleShortenUrl = ({ fetchUrls }) => {
 
   useEffect(() => {
     getOriginalLinkBySlug();
-  }, []);
+  }, [slug]);
   return <></>;
 };
 export default HandleShortenUrl;
