+++ tailwind.config.js
@@ -21,6 +21,7 @@ module.exports = {
         "new-light-gray": "#f0f0f0",
         "light-olive-green": "#0ead7a",
         "olive-green-hover": "#49bf99",
+        "extra-light-olive": "#2bed8f",
       },
       boxShadow: {
         "custom-box-shadow": "10px 10px 5px 200px rgba(0,0,0,1)",
